requirements.txt
requests
datetime
tweepy
gunicorn==19.7.1
pillow